requirements.txt
# Core (install with: pip install -r requirements.txt)
mcp
python-dotenv

# Optional: drdroid-debug-toolkit (for drd-backed MCP servers)
# git+https://github.com/DrDroidLab/drdroid-debug-toolkit.git@master
